+++ include/pangolin/xml/rapidxml_utils.hpp
@@ -37,7 +37,7 @@ namespace rapidxml
             
             // Determine stream size
             stream.seekg(0, ios::end);
-            size_t size = stream.tellg();
+            size_t size = (size_t)stream.tellg();
             stream.seekg(0);   
             
             // Load data and add terminating 0
